refactor: flatten the package, making repo root the wsstat package
